Extract BuildingModuleGraph implementation
